root ts files native-web split
