+++ packages/react-native-reanimated/src/UpdateLayoutAnimations.ts
@@ -1,39 +1,8 @@
 'use strict';
-import { createSerializable } from 'react-native-worklets';
-
-import { SHOULD_BE_USE_WEB } from './common';
 import type {
-  LayoutAnimationBatchItem,
   LayoutAnimationFunction,
   LayoutAnimationType,
 } from './commonTypes';
-import { configureLayoutAnimationBatch } from './core';
-
-function createUpdateManager() {
-  const animations: LayoutAnimationBatchItem[] = [];
-  // When a stack is rerendered we reconfigure all the shared elements.
-  // To do that we want them to appear in our batch in the correct order,
-  // so we defer some of the updates to appear at the end of the batch.
-  const deferredAnimations: LayoutAnimationBatchItem[] = [];
-
-  return {
-    update(batchItem: LayoutAnimationBatchItem, isUnmounting?: boolean) {
-      if (isUnmounting) {
-        deferredAnimations.push(batchItem);
-      } else {
-        animations.push(batchItem);
-      }
-      if (animations.length + deferredAnimations.length === 1) {
-        this.flush();
-      }
-    },
-    flush(this: void) {
-      configureLayoutAnimationBatch(animations.concat(deferredAnimations));
-      animations.length = 0;
-      deferredAnimations.length = 0;
-    },
-  };
-}
 
 /**
  * Lets you update the current configuration of the layout animation or shared
@@ -51,35 +20,12 @@ function createUpdateManager() {
  *   those that were updated later). This is used to retain the correct ordering
  *   of shared elements. Defaults to `false`.
  */
-export let updateLayoutAnimations: (
+export const updateLayoutAnimations: (
   viewTag: number,
   type: LayoutAnimationType,
   config?: Keyframe | LayoutAnimationFunction,
   isUnmounting?: boolean,
   sharedTransitionTag?: string
-) => void;
-
-// is-tree-shakable-suppress
-if (SHOULD_BE_USE_WEB) {
-  updateLayoutAnimations = () => {
-    // no-op
-  };
-} else {
-  const updateLayoutAnimationsManager = createUpdateManager();
-  updateLayoutAnimations = (
-    viewTag,
-    type,
-    config,
-    isUnmounting,
-    sharedTransitionTag
-  ) =>
-    updateLayoutAnimationsManager.update(
-      {
-        viewTag,
-        type,
-        config: config ? createSerializable(config) : undefined,
-        sharedTransitionTag,
-      },
-      isUnmounting
-    );
-}
+) => void = () => {
+  // no-op
+};
