+++ packages/react-native-reanimated/src/commonConfigHelper.ts
@@ -1,0 +1,13 @@
+'use strict';
+
+/** @deprecated This function is a no-op in Reanimated 4. */
+export function addWhitelistedNativeProps(
+  _props: Record<string, boolean>
+): void {
+  // Do nothing. This is just for backward compatibility.
+}
+
+/** @deprecated This function is a no-op in Reanimated 4. */
+export function addWhitelistedUIProps(_props: Record<string, boolean>): void {
+  // Do nothing. This is just for backward compatibility.
+}
