+++ packages/react-native-reanimated/src/mappers.native.ts
@@ -109,11 +109,21 @@ function createMapperRegistry() {
     }
   }
 
-  const schedulingFunction = requestAnimationFrame;
+  const schedulingFunction = IS_JEST
+    ? requestAnimationFrame
+    : globalThis.requestAnimationFrameFinalizer;
 
   function scheduledMapperRun() {
     runRequested = false;
     mapperRun();
+    if (!IS_JEST) {
+      // We always run mappers on native.
+      schedulingFunction(scheduledMapperRun);
+    }
+  }
+
+  if (!IS_JEST) {
+    schedulingFunction(scheduledMapperRun);
   }
 
   global.__mapperRun = mapperRun;
@@ -194,10 +204,14 @@ function createMapperRegistry() {
         sv.addListener(mapper.id, () => {
           mapper.dirty = true;
           isAnyMapperDirty = true;
-          maybeRequestUpdates();
+          if (IS_JEST) {
+            maybeRequestUpdates();
+          }
         });
       }
-      maybeRequestUpdates();
+      if (IS_JEST) {
+        maybeRequestUpdates();
+      }
     },
     stop: (mapperID: number) => {
       const mapper = mappers.get(mapperID);
